+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/interop/InteropAssertionsTest.java
@@ -59,6 +59,7 @@ import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.source.Source;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.api.test.polyglot.ProxyLanguage;
+import com.oracle.truffle.api.utilities.TriState;
 
 @SuppressWarnings("deprecation")
 public class InteropAssertionsTest extends InteropLibraryBaseTest {
@@ -401,6 +402,98 @@ public class InteropAssertionsTest extends InteropLibraryBaseTest {
         assertTrue(l.isMetaObject(v));
         assertSame(testQualifiedName, l.getMetaQualifiedName(v));
         assertSame(testSimpleName, l.getMetaSimpleName(v));
+    }
+
+    @FunctionalInterface
+    interface IsSameOrUndefined {
+
+        TriState isSameOrUndefined(Object receiver, Object other);
+
+    }
+
+    @FunctionalInterface
+    interface IdentityHashCode {
+
+        int identityHashCode(Object receiver) throws UnsupportedMessageException;
+
+    }
+
+    @ExportLibrary(InteropLibrary.class)
+    static class IdentityTest implements TruffleObject {
+
+        IsSameOrUndefined isSameOrUndefined;
+        IdentityHashCode identityHashCode;
+
+        @ExportMessage
+        TriState isSameOrUndefined(Object other) {
+            if (isSameOrUndefined == null) {
+                return TriState.UNDEFINED;
+            }
+            return isSameOrUndefined.isSameOrUndefined(this, other);
+        }
+
+        @ExportMessage
+        int identityHashCode() throws UnsupportedMessageException {
+            if (identityHashCode == null) {
+                throw UnsupportedMessageException.create();
+            }
+            return identityHashCode.identityHashCode(this);
+        }
+
+    }
+
+    @Test
+    public void testIsSame() {
+        IdentityTest v0 = new IdentityTest();
+        IdentityTest v1 = new IdentityTest();
+        InteropLibrary l0 = createLibrary(InteropLibrary.class, v0);
+        InteropLibrary l1 = createLibrary(InteropLibrary.class, v1);
+
+        // correct usage
+        v0.isSameOrUndefined = (r, o) -> TriState.valueOf(o == v0);
+        v0.identityHashCode = (r) -> System.identityHashCode(r);
+        v1.isSameOrUndefined = (r, o) -> TriState.valueOf(o == v1);
+        v1.identityHashCode = (r) -> System.identityHashCode(r);
+        assertTrue(l0.isSame(v0, v0, l0));
+        assertTrue(l1.isSame(v1, v1, l1));
+        assertFalse(l0.isSame(v0, v1, l1));
+
+        // missing identity hash code
+        v0.isSameOrUndefined = (r, o) -> TriState.valueOf(o == v0);
+        v0.identityHashCode = null;
+        assertFails(() -> l0.isSame(v0, v1, l1), AssertionError.class);
+
+        // symmetry violated
+        v0.isSameOrUndefined = (r, o) -> {
+            if (o == v1) {
+                return TriState.TRUE;
+            }
+            return TriState.UNDEFINED;
+        };
+        v0.identityHashCode = (r) -> System.identityHashCode(r);
+        assertFails(() -> l0.isSame(v0, v1, l1), AssertionError.class);
+
+        // reflexivity violated
+        v0.isSameOrUndefined = (r, o) -> {
+            if (o == v0) {
+                return TriState.FALSE;
+            }
+            return TriState.UNDEFINED;
+        };
+        v0.identityHashCode = (r) -> System.identityHashCode(r);
+        assertFails(() -> l0.isSame(v0, v0, l0), AssertionError.class);
+
+        // invalid identity hash code
+        v0.isSameOrUndefined = (r, o) -> TriState.valueOf(o == v0 || o == v1);
+        v0.identityHashCode = (r) -> 42;
+
+        v1.isSameOrUndefined = (r, o) -> TriState.valueOf(o == v0 || o == v1);
+        v1.identityHashCode = (r) -> 43;
+        assertFails(() -> l0.isSame(v0, v1, l1), AssertionError.class);
+
+        // fix invalid identity hash code
+        v1.identityHashCode = (r) -> 42;
+        assertTrue(l0.isSame(v0, v1, l1));
     }
 
 }