+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotList.java
@@ -130,12 +130,16 @@ class PolyglotList<T> extends AbstractList<T> implements HostWrapper {
 
     @Override
     public int hashCode() {
-        return HostWrapper.hashCode(this);
+        return HostWrapper.hashCode(languageContext, guestObject);
     }
 
     @Override
     public boolean equals(Object o) {
-        return HostWrapper.equals(this, o);
+        if (o instanceof PolyglotList) {
+            return HostWrapper.equals(languageContext, guestObject, ((PolyglotList<?>) o).guestObject);
+        } else {
+            return false;
+        }
     }
 
     static final class Cache {