+++ truffle/src/com.oracle.truffle.sl/src/com/oracle/truffle/sl/runtime/SLFunction.java
@@ -44,12 +44,13 @@ import java.util.logging.Level;
 
 import com.oracle.truffle.api.Assumption;
 import com.oracle.truffle.api.CallTarget;
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLogger;
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Cached;
+import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.ReportPolymorphism;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.interop.InteropLibrary;
@@ -60,6 +61,7 @@ import com.oracle.truffle.api.nodes.DirectCallNode;
 import com.oracle.truffle.api.nodes.IndirectCallNode;
 import com.oracle.truffle.api.source.SourceSection;
 import com.oracle.truffle.api.utilities.CyclicAssumption;
+import com.oracle.truffle.api.utilities.TriState;
 import com.oracle.truffle.sl.SLLanguage;
 import com.oracle.truffle.sl.nodes.SLUndefinedFunctionRootNode;
 
@@ -183,6 +185,28 @@ public final class SLFunction implements TruffleObject {
     }
 
     @ExportMessage
+    @SuppressWarnings("unused")
+    static final class IsSameOrUndefined {
+        @Specialization
+        static TriState doSLFunction(SLFunction receiver, SLFunction other) {
+            /*
+             * SLFunctions are potentially identical to other SLFunctions.
+             */
+            return receiver == other ? TriState.TRUE : TriState.FALSE;
+        }
+
+        @Fallback
+        static TriState doOther(SLFunction receiver, Object other) {
+            return TriState.UNDEFINED;
+        }
+    }
+
+    @ExportMessage
+    static int identityHashCode(SLFunction receiver) {
+        return System.identityHashCode(receiver);
+    }
+
+    @ExportMessage
     Object toDisplayString(@SuppressWarnings("unused") boolean allowSideEffects) {
         return name;
     }
@@ -190,7 +214,7 @@ public final class SLFunction implements TruffleObject {
     /**
      * We allow languages to execute this function. We implement the interop execute message that
      * forwards to a function dispatch.
-     * 
+     *
      * Since invocations are potentially expensive (result in an indirect call, which is expensive
      * by itself but also limits function inlining which can hinder other optimisations) if the node
      * turns megamorphic (i.e. cache limit is exceeded) we annotate it with {@ReportPolymorphism}.